Update s3.yml
In this template:

AWSTemplateFormatVersion specifies the CloudFormation template version.
Description provides a brief description of the template's purpose.
Under Resources, you define the AWS resources you want to create. In this case, we're creating an S3 bucket with the logical ID MyS3Bucket.
Type indicates the resource type, which is AWS::S3::Bucket for an S3 bucket.
Under Properties, you can specify various properties for the resource. In this example, we set the bucket name to my-unique-bucket-name and its access control to Private.
Make sure to replace my-unique-bucket-name with a globally unique name for your S3 bucket.

Remember that this is a basic example. Depending on your use case, you might want to include additional properties or configurations. You can refer to the AWS CloudFormation documentation for AWS::S3::Bucket for more information about the available properties and options.

To deploy this template, you can use the AWS CLI or any other method of your choice to create a CloudFormation stack using the template file.